small comments fix

+++ test_vis.py
@@ -116,7 +116,7 @@ def test_sky_illumination(input_dem_path, output_path):
     y_res = -t[4]
     input_dem_arr = input_dem_dataset.read()[0]
     skyillumination_arr = rvt.vis.sky_illumination(dem=input_dem_arr, resolution=x_res)
-    skyillumination_arr = skyillumination_arr.astype('float64')
+    skyillumination_arr = skyillumination_arr.astype('float32')
     profile = input_dem_dataset.profile
     profile.update(dtype='float32')
     output_skyilum = rio.open(output_path, "w", **profile)
